requirements.txt
Django==5.0.6
whitenoise
gunicorn
psycopg2-binary
dj-database-url
pytz
djangorestframework>=3.0,<4.0